requirements.txt
# Project 5001 Dependencies
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
mutagen>=1.47.0

# Optional: For enhanced audio processing
# eyed3>=0.9.7  # Alternative to mutagen for MP3 tagging

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0 